+++ src/Serivces/Stock/Stock.API/SyncDataServices/Soap/ValuteDataService.cs
@@ -11,7 +11,7 @@
         /// <param name="amount">Сумма для конвертации.</param>
         /// <param name="currencyFrom">Код валюты исходной суммы amount.</param>
         /// <param name="currencyTo">Код валюты, в которую нужно конвертировать.</param>
-        Task<double> ConvertValute(double amount, string currencyFrom, string currencyTo);
+        Task<decimal> ConvertValute(decimal amount, string currencyFrom, string currencyTo);
     }
 
     /// <summary>
@@ -19,6 +19,9 @@
     /// </summary>
     public class ValuteDataService : IValuteDataService
     {
+        //TODO: возможны проблемы с Hot-Reaload. При его использовании запросы к сервису 
+        // падают с исключением.
+        // Временное решение - перезапустить приложение.
         private readonly ICentralBankService _centralBankService;
 
         public ValuteDataService(ICentralBankService centralBankService)
@@ -27,7 +30,7 @@
         }
 
         /// <inheritdoc/>
-        public async Task<double> ConvertValute(double amount, string currencyFrom, string currencyTo)
+        public async Task<decimal> ConvertValute(decimal amount, string currencyFrom, string currencyTo)
         {
             // Если конвертация бессмыслена
             if (currencyFrom.Equals(currencyTo))
@@ -36,8 +39,8 @@
             currencyFrom = currencyFrom.ToUpperInvariant();
             currencyTo = currencyTo.ToUpperInvariant();
 
-            double amountRub = amount, // сумма в рублях
-                   amountCurrencyTo; // сумма в новой валюте
+            decimal amountRub = amount, // сумма в рублях
+                    amountCurrencyTo; // сумма в новой валюте
 
             // Список курсов валют на сегодня
             var cursOnDate = await _centralBankService.GetCursOnDateXMLAsync(DateTime.Now);
@@ -60,10 +63,10 @@
 
             return amountCurrencyTo;
 
-            double ConvertToRUB(double amount, ValuteCursOnDate cursTo) => amount *
-                                                                           (double)(cursTo.Vcurs / cursTo.Vnom);
+            decimal ConvertToRUB(decimal amount, ValuteCursOnDate cursTo) => amount *
+                                                                             (cursTo.Vcurs / cursTo.Vnom);
 
-            double ConvertFromRubToCurrency(double amountRub, ValuteCursOnDate cursTo) => amountRub / (double)(cursTo.Vcurs / cursTo.Vnom);
+            decimal ConvertFromRubToCurrency(decimal amountRub, ValuteCursOnDate cursTo) => amountRub / (cursTo.Vcurs / cursTo.Vnom);
         }
     }
 }
